Updating GameAPI

+++ Game.API/Domain/Services/GameRepository.cs
@@ -13,7 +13,7 @@ namespace Game.API.Domain.Services
         public GameRepository(GamesContext context) => _context = context;
 
         public IEnumerable<Games> GetGames() => _context.Games;
-        public Games GetGame(int gameId) => _context.Games.Where(x => x.Id == gameId).Single();
+        public Games GetGame(long gameId) => _context.Games.Where(x => x.Id == gameId).Single();
         public IEnumerable<GameRatings> GetGameRatings() => _context.GameRatings;
         public void SaveDetail(Games game)
         {
