+++ Game.API/Domain/Services/IGameRepository.cs
@@ -6,7 +6,7 @@ namespace Game.API.Domain.Services
     public interface IGameRepository
     {
         IEnumerable<Games> GetGames();
-        Games GetGame(int gameId);
+        Games GetGame(long gameId);
         void SaveDetail(Games game);
         IEnumerable<GameRatings> GetGameRatings();
     }
